+++ examples/simple-subscriptions/index.ts
@@ -1,67 +1,32 @@
 import "reflect-metadata";
 import http from "node:http";
 import path from "node:path";
-import { ApolloServer } from "@apollo/server";
-import { expressMiddleware } from "@apollo/server/express4";
-import { ApolloServerPluginDrainHttpServer } from "@apollo/server/plugin/drainHttpServer";
-import { ApolloServerPluginLandingPageLocalDefault } from "@apollo/server/plugin/landingPage/default";
-import bodyParser from "body-parser";
-import cors from "cors";
-import express from "express";
-import { useServer } from "graphql-ws/lib/use/ws";
+import { createYoga } from "graphql-yoga";
 import { buildSchema } from "type-graphql";
-import { WebSocketServer } from "ws";
-import { SampleResolver } from "./resolver";
+import { NotificationResolver } from "./notification.resolver";
+import { pubSub } from "./pubsub";
 
 async function bootstrap() {
   // Build TypeGraphQL executable schema
   const schema = await buildSchema({
     // Array of resolvers
-    resolvers: [SampleResolver],
+    resolvers: [NotificationResolver],
     // Create 'schema.graphql' file with schema definition in current directory
     emitSchemaFile: path.resolve(__dirname, "schema.graphql"),
+    // Publish/Subscribe
+    pubSub,
   });
-
-  // Create an Express app and HTTP server
-  // The WebSocket server and the ApolloServer will be attached to this HTTP server
-  const app = express();
-  const httpServer = http.createServer(app);
-
-  // Create WebSocket server using the HTTP server
-  const wsServer = new WebSocketServer({
-    server: httpServer,
-    path: "/graphql",
-  });
-  // Save the returned server's info so it can be shutdown later
-  const serverCleanup = useServer({ schema }, wsServer);
 
   // Create GraphQL server
-  const server = new ApolloServer({
+  const yoga = createYoga({
     schema,
-    csrfPrevention: true,
-    cache: "bounded",
-    plugins: [
-      // Proper shutdown for the HTTP server
-      ApolloServerPluginDrainHttpServer({ httpServer }),
-      // Proper shutdown for the WebSocket server
-      {
-        async serverWillStart() {
-          return {
-            async drainServer() {
-              await serverCleanup.dispose();
-            },
-          };
-        },
-      },
-      ApolloServerPluginLandingPageLocalDefault({ embed: true }),
-    ],
+    graphqlEndpoint: "/graphql",
   });
 
-  // Start server
-  await server.start();
-  app.use("/graphql", cors<cors.CorsRequest>(), bodyParser.json(), expressMiddleware(server));
+  // Create server
+  const httpServer = http.createServer(yoga);
 
-  // Now that the HTTP server is fully set up, we can listen to it
+  // Start server
   httpServer.listen(4000, () => {
     console.log(`GraphQL server ready at http://localhost:4000/graphql`);
   });